Change amount field to type DecimalField

+++ Server/users/models.py
@@ -5,7 +5,7 @@ from datetime import date
 
 class CaffeineIntake(models.Model):
     caffeine = models.PositiveSmallIntegerField(default=0)
-    amount = models.PositiveSmallIntegerField(default=0)
+    amount = models.DecimalField(default=0, max_digits=7, decimal_places=2)
     type = models.CharField(max_length=100)
     measurement = models.CharField(max_length=100)
     date = models.DateField(default=date.today)
